# openapi.yaml
openapi: 3.2.0
info:
  title: GREENHERB API
  description: Plataforma Web de Gestão Inteligente de uma Estufa
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "409":
          description: Conflict
  /auth/login:
    post:
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
  /auth/refresh:
    post:
      tags:
        - Auth
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
...
  /ervas:
    get:
      tags:
        - Ervas Aromáticas
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
    post:
      tags:
        - Ervas Aromáticas
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErvaAromatica"
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /ervas/importar:
    post:
      tags:
        - Ervas Aromáticas
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Ficheiro CSV com Ervas a importar
      responses:
        "201":
          description: Created
        "207":
          description: Multi-Status (Erros parciais)
        "400":
          description: Bad Request
  /ervas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Ervas Aromáticas
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
    patch:
      tags:
        - Ervas Aromáticas
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErvaAromatica"
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
    delete:
      tags:
        - Ervas Aromáticas
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK - Soft Delete
        "404":
          description: Not Found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access
  schemas:
    UserRegister:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
    ErvaAromatica:
      type: object
      required: [nome, nomeCientifico, variedade, familia, cicloDiasFim]
      properties:
        nome:
          type: string
        nomeCientifico:
          type: string
        variedade:
          type: string
        familia:
          type: string
        cicloDiasFim:
          type: number
        descricao:
          type: string
        ativo:
          type: boolean
...